ml_pipeline/requirements-ml.txt
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.66.0
numpy>=1.26.0
pillow>=10.0.0
timm>=0.9.8
